requirements.txt
Flask==2.2.2
loguru==0.6.0
openai==0.27.0
slackclient==2.9.4
slackeventsapi==3.0.1
